Refactor the disassembler code
