+++ smwe_rom/src/disassembler/mod.rs
@@ -1,1 +1,6 @@
-pub mod instructions;
+// A lot of the disassembler code is "borrowed" from DiztinGUIsh, an SNES ROM disassembler and debugger written in C#.
+// https://github.com/Dotsarecool/DiztinGUIsh
+
+pub mod instruction;
+pub mod opcodes;
+pub mod rom_disassembly;
